// src/components/container/Formily/Models/Widget/input.ts
export type ValidatorInterface = {
	strategy: string;
	triggerType: string;
	driveList: string[];
	rangeRuleList: {
		fieldName: string;
		condition: string;
		target: number;
		unit: string;
		message: string;
		messageLangKey: string;
	}[];
	validator: string;
	message: string;
	messageLangKey: string;
	format: string | undefined;
	pattern: string;
	len: number | null;
	max: number | null;
	min: number | null;
	exclusiveMaximum: number | null;
	exclusiveMinimum: number | null;
	whitespace: boolean;
};